Changes to TA2CalArray (default NaI class) and updated MC config files
A global time shift was applied to ReadDecoded (MC handler) in TA2CalArray (default NaI class). This time shift of -1.6 ns caused a rejection of roughly 50 percent of NaI hits for MC data using the default setup files since the TDC cut was set to accept 0-200. I have fixed this in two ways, an wider window is now included on the NaI config file, and the time shift was removed. I have also adapted the MWPC and PID config files. The PID file was chosen from the G4 release package and has been checked. The MWPC file was provided by A. Mushkarenkov.