New system Histograms included in output Tree File.
New histogram output from TA2GoAT, used for fast diagnostics of file issues. These histogram include detector-veto correlations (dE_E and phi correlations) for CB and TAPS, as well as the hits, adc, and tdc patterns as a function of event number. To add the adc and tdc patterns properly, the issues with the Trigger readout from real data is now fixed as well. This was done to reject the TAPS pulser from the ADC/TDC hit patterns. This means that multiplicities in the trigger tree are also fixed.