Correction of inconsistant NULL markers
Each apparatus/detector in Acqu uses some inconsistant NULL markers to signify that no hit or energy deposit was found for a given track. These values are some large +ve or -ve large values (greater than e6). It is fully appropriate to store a value of 0.0 for an energy deposition (instead of +1e9) when no deposition is recorded. This standard has now been applied to data stored in the output tree of TA2GoAT.