Finish macro for multi-thread Acqu Analysis
Finish macro which appends input file name to the output file. This is primarily helpful for analysis with Worker. An output directory can be specified, if none is specified the global output directory (defined in AR.dat) is used. If this is also unspecified, the users home directory is used as a default.