Core: Improve End-Of-Run handling
Copied from Elog entry 18054:
Changes to TDAQexperiment:
Moved the block of code where a slave system checks the event ID to just after the ADC read loop and just before
the scaler-read code block. This means that the slave system knows if the master is signaling end-of-run before
it starts to read the scalers. Thus a scaler read is performed on all systems (which have scalers) when the
end-of-run is issued from the DAQ gui. Previously only the master was reading scalers at end-of-run
Checked that slaves do scaler read at end-of-run on the debugger...OK
Also ran the event-merger so that all data streams are dumped to disk individually.
Visually check the individual stream files...e.g.
> od -d scratch/tagga_0.dat | less
> /65278
65278 = 0xfefe is the scaler buffer marker
See scaler blocks in the data at events 9999, 19999.... and also at the last event before end-of-run
This side appears to be working.

However the last scaler read was not showing up in merged data, and several events at the end of the
individual-stream files were also not showing up in the merged data file.
Traced the problem to TA2DataServer::MultiProcess(). When the end-buffer is detected coming from the daq master,
there can still be a merged data buffer (containing the last few events) sitting in the system. It has not been
written to file as the buffer is not full. Now when end-buffer is detected from the master, TA2DataServer checks
the merged-data buffer and if there is anything there it writes it to the merged output file, BEFORE writing the
end-of-file buffer.
Note that this last partially-filled buffer will probably not be caught by the online data sorter. When running
online not all merged data buffers are analysed as this could slow down data collection.
Checked the merged files for a sequence of run/end commands from the gui. Each file contains scaler-read blocks
from exp-trigger, beampolmon and tagga at events 9999, .....
Each also contains a scaler block from exp-trigger, beampolmon, tagga in the last event.

AcquRoot seems to process files CBTagg_3085/6/7.dat correctly offline. These files were taken in one daq session
(without taking the daq down in between files)
It seems to be working correctly, but needs a bit more testing.