Core: Direct File IO implementation, see notes below before use!
Activate with UseDirectIO: statement in AR.dat in --batch mode.
You may need to recompile ROOT with proper lzma support
(not the ROOT builtin support by $ROOTSYS/lib/liblzma.a)!

It circumvents the complex thread mechanism and feeds the Analysis directly
with chunks from a file. It supports uncompressed files as well
as *.xz compressed ones.

Note: This is highly untested and is only for testing
the impact on machines like blaster.