User: PerRunCorr: Add option header line to parameter files, i.e.:
You can optionally specify the operation how the correction parameters
shall be applied, for example put in the first line of cb_t0.txt

"# applywith=add"

adds the subsequent given parameters to the values read from the Calib
database.

Note: The line must be the first and must start with an "#"

Default is "applywith=multiply", if the line is omitted completely.